backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pytesseract==0.3.13
Pillow==11.1.0
opencv-python-headless==4.11.0.86
httpx==0.28.1
python-multipart==0.0.20
imagehash==4.3.1
pybktree==1.1
